Edge Exclusion for pathTo function
You can now exclude certain edges in your pathTo function with the
paramater `only` that expects a function that takes an edge as an
argument and returns true if and only if the edge should be used
for the path.

Use Case:
Only include edges on the path that have a rating of x or higher.