issue #224: allow optional content-id for batches
this allows clients to send a content-id for each batch part
the server will return the content-id in each part response so the client can more easily map requests and responses

the content-id will not be validated or checked for uniqueness by the server. it is up to the client to send sensible content-ids to the server
currently, the server will not store the content-ids anywhere. they are just used within one batch request and then "forgotten"