add words to fulltext index
this parses the fulltext-indexed attributes of documents when there's a fulltext index, and adds the individual words to the index.
As the fulltext index is case-sensitive, all words are added to the index in lower case.
The text tokenisation implementation is still very naive and currently works properly only for character ranges [a-z] and [A-Z].
Unicode words are also supported, but they are not normalised nor lower-cased yet. Additionally, unicode punctuation characters are not excluded and will also be added to the index.
Updating documents that are fulltext-indexed currently does not work.