[Small improvements] In the function ResizeAssociativeArray(), array->_table is already initialized to '\0's, so there is no need to call clearElement() for each element. An associative array is only used by PQIndex and MasterTable_t. For PQIndex the function ClearElementPQIndex() will be used, which sets the whole element to '\0's. For MasterTable_t the function tablePositionClearElement() will be used, which sets all but one field to 0/NULL (the only field which is not set is _vectorNum). If the clearElement() should be called, maybe TRI_Allocate() could be called with false, to avoid double initialization. When copying elements, the for loop stops when the number of elements is the same as the original number of elements.
In the function TRI_FindByKeyAssociativeArray(), there is no need to check whether the element is not empty and is equal key element, only if it is not empty (as it is done in the function TRI_InsertKeyAssociativeArray()).
In the function TRI_FindByElementAssociativeArray(), there is no need to check whether the element is not empty and is equal element element, only if it is not empty (as it is done in the function TRI_InsertElementAssociativeArray()).