added option `--server.disable-authentication-unix-sockets`
with this option, authentication can be disabled for all requests coming
in via UNIX domain sockets, enabling clients located on the same host as
the ArangoDB server to connect without authentication.
Other connections (e.g. TCP/IP) are not affected by this option.

The default value for this option is `false`.
Note: this option is only supported on platforms that support Unix domain
sockets.