Reorganise unit tests.
In JS framework we now find the available test files by directory
listing. This means one can simply drop a file in the right
directory to add a test.

We use a naming convention for filenames which is explained
in the README files in the corresponding directories.

Old unit tests do not yet use this but do use the new names and
should still work as usual.