Fix and unify asterisks, tested with GitBook.
Italic asterisks should be done by escaping the asterisk and wrapping it with html i-tag. It seems to be the only safe technique (3 asterisks to display 1 italic asterisk in a list works, unless followed by more markup for italic text; 1 escaped asterisk wrapped by i-tag works regardless of the rest of the line).