Re-implemented require/module based on Node 4.1.0
The require function should now behave as expected and treat ArangoDB's module paths as sources for "external" modules.
The "assert", "buffer" and "path" modules have been moved out of ./js/node into the bootstrap modules.
The module "org/arangodb/foxx/arangoApp" (as well as any modules depending on it) is incompatible with the new module loading behaviour and will be have to be rewritten from scratch.
The "Package" class has been removed entirely and is no longer needed.
Support for the database-stored modules is currently missing and needs to be re-implemented within the module loader.