Make sure constructor is called in legacy-extended classes
This is a horrible workaround for the lack of a spread operator in the V8 version we are currently using.
The super keyword has no way to pass a dynamic number of arguments without using the spread operator
but someone might reasonably want to use additional arguments in a regular subclass that is then
legacy-extended with a constructor property. This should at least protect against the majority of such
cases even if it creates a silly function signature.

See #1592.