Fix a use after free for TRI_vocbase_t objects.
If a background task was scheduled for a database for later
execution and the database was dropped in the meantime and
if there was a shortage of V8 contexts, then the database
could have been dropped before the job began.
The bug was that V8Job received a (used) vocbase, but only
increased the refcount when a V8 context was entered. This
could be much later and so the database could be gone.