Resolves issue to allow cluster statistics with https and executes a series of promises to be resolved for each coordinator
Problem: 

 ```2016-09-20T22:24:14Z [5510] ERROR SyntaxError: Unexpected token u in JSON at position 0
2016-09-20T22:24:14Z [5510] ERROR     at Object.parse (native)
2016-09-20T22:24:14Z [5510] ERROR     at coordinators.map.e (/usr/share/arangodb3/js/apps/system/_admin/aardvark/APP/statistics.js:504:19)
2016-09-20T22:24:14Z [5510] ERROR     at Array.map (native)
2016-09-20T22:24:14Z [5510] ERROR     at Route._handler (/usr/share/arangodb3/js/apps/system/_admin/aardvark/APP/statistics.js:499:39)

Solution:
The Internal.download function should be treated as a callback rather than a synchronous function. I'd recommend an implementation of http.get or request instead if the above download function isn't changed. Because the download function is synchronous it sometimes returns undefined, which continuously throws an error in the JSON.parse function. 

Additionally, I added ssl to the string replace function and added additional error handling for when there is an error into the merge history.