fix upgrade/install foxx in cluster environment
When install/upgrade, arangodb response is "No such file or directory".
This is because serviveInfo path already join with temporary Path, and then
in "uploadToPeerCoordinators" it join twice with temporary path.
Another problem, arangodb upgrade url is "update" not "upgrade".
Last problem, when Sync foxx with another cluster, use "coordinatorTransactionID"
but there is no that key response from ArangoClusterComm.