Adding note about configuration file and endpoint.
Took me a while to find how to change server.endpoint. When editing provided plist file and using `brew services start arangodb` `ProgramArguments` from plist are ignored for some reason. So the only way to switch from `127.0.0.1:8529` to `0.0.0.0:8529` was to edit `arangod.conf` file and restart services with brew. After that I was able to access web interface remotely.