Revert "switch immediately to async writes when socket would block during write"
This reverts commit 96067a9315edbf2efa87f75f2a2faa677d6051d4.

The change resulted in a stackoverflow instead of a "hanging" connection
we might need to use something like TCP Keep Alive to handle
disconnected clients. The bug we are currently working on might not be
fixed by TCP Keep Alive as we are in a heavy load situation. Still the
socket should not block during write. It could do so if the client does
not receive the data form the network and the server side networkbuffer
runs full at some point. There might be other situations we are not
aware of.