Fix synchronous replication.
It was possible that a follower got into sync after the list of
followers was retrieved, but before a single document was written.
This dropped this document for the new follower because the leader
forgot to replicate it. This lead to document loss in case of a
failover or planned leader change.