Let performRequests optionally ignore 404 coll not found.
This is important in certain cases, in which something is sent to all
shard leaders, say, and the coordinator might have an outdated
information about who a shard leader is. In that case it can happen
that the old shard leader returns 404 coll not found, and this must then
be treated as a retry after loading the latest info from the agency.