implement heap merge for cluster blocks
This reduces the amount of comparisons in a
gather block from 560k to 370k. The collection
used consists of 9 shards containing 400k
documents. The amount of returned documents
was 70k docs.

DO ONLY MERGE WHEN ALL TESTS HAVE PASSED!

Cluster tests are red for me at positions
that seem to be unrelated.