honor specified COLLECT OPTIONS
for example, when the user explicitly asks for the COLLECT method
to be `sorted`, the optimizer will now not produce an alternative
version of the plan using the hash method.

additionally, if the user explcitly asks for the COLLECT method to
be `hash`, the optimizer will now change the existing plan to use
the hash method if possible instead of just creating an alternative
plan.

`COLLECT ... OPTIONS { method: 'sorted' }` => always use sorted method
`COLLECT ... OPTIONS { method: 'hash' }`   => use hash if this is technically possible
`COLLECT ...` (no options)                 => create a plan using sorted, and another plan using hash method