Avoid mutex in ReadWriteLock where possible.
To compare the performance of the old and new implementations I ran a number of synthetic benchmarks for different scenarios (varying number of threads, reader/writer ratio and time a lock is held) 
on 4 different machines with up to 4 sockets and 160 hardware threads. Over all scenarios, the mean throughput of new implementation was ~7% higher. A paired Wilcoxon test also confirmed that the throughput of the new implementation is significantly higher (p = 2.502e-06).