New ReadWriteSpinLock implementation.
To compare the performance of both implementations I ran a number of synthetic benchmarks for different scenarios (varying number of threads, reader/writer ratio and time a lock is held) on 4 different machines with up to 4 sockets and 160 hardware threads. Over all scenarios, the mean throughput of new implementation was slightly higher (~2%), but not significantly. However, the memory requirements of the new implementation are significantly lower (4 bytes vs 8360 bytes with the default of 64 stripes).