make arangorestore not report errors about the target database not found
when it is invoked with option `--create-database true`. In this case,
arangorestore should not print an error but simply create the target
database. This patch implements this change.

Additionally, it aborts the startup of arangorestore in case there are
any errors. Previously, the startup phase went on even when no connection
could be established, which led to useless follow-up errors being printed.