abort startup when using SSLv2 for a server endpoint
or when connecting with a client tool via an SSLv2 connection.

SSLv2 has been disabled in the OpenSSL library by default in recent versions
because of security vulnerabilities inherent in this protocol.

This also fixes a compile error when having installed a recent
version of OpenSSL on the system which does not provide SSLv2 at all.