defer creation of ArangoSearch threadpool
moved threadpool creation from IResearchFeature constructor into
method IReseachFeature::initializeAsync(). This prevents the
threadpool from being created when the IResearchFeature object is
created.
In arangod, the feature object is created at the very start, before
it is clear if we actually need the IResearchFeature on the server.
In case we are an agency node, we will not need the IResearchFeature
at all, so we will disable it at some point during the startup
sequence. This however makes more sense if the feature hasn't yet
started a threadpool.

IResearchFeature::initializeAsync isn't private because it needs to
be called from the unit tests to properly initialize the async tests
for the IResearchFeature