centralized more of the sharding strategies code
in arangod/Sharding/ShardingStrategiesDefault.*
The sharding strategy "enterprise-compat" is now also available
in the community version so users can switch between enterprise
and community version and vice versa without incompatibilities.

There are mock sharding strategy implementations of the enterprise
sharding strategies for smart edge collections now in the community
edition too. These do not actually shard any data, but they allow
increase the compatibility between enterprise and community edition
in case a downgrade is done from enterprise to community while
using smart edge collections.
When doing that, the previous smart edge collection will now have
a mock smart-edge collection sharding strategy that will always
throw an exception when actually asked for the location of a document.
The exception will indicate that the sharding strategy is only
available in the enterprise edition. This effectively disables
accessing that particular smart edge collection in the community
edition, but this is unavoidable given the fact that there is no
smart edge collection code in that version at all.
As long as the collection is not used, this should at least allow
using the rest of the cluster.