add script counting network connections (#6754)
This was added (and is currently not being used) to find a problem where a server in shutdown would hammer the agency with new connections. It was decided to implement a corresponding monitoring directly in the monitoring system (ganglia) rather than in the context of `testing.js`. Nevertheless, the script is retained for posteriority here.