A code worth making a project
The puzzle, also called the Tower of Brahama or End of the World Puzzle, was created out of an old Hindu legend. The story tells of a Hindu temple at the center of the world where priests were given a stack of 64 golden disks of decreasing size. The disks were stacked on one of three towers.