do not create `_routing` collection for new installations/new databases,
as it is not needed anymore. Redirects to the web interface's login screen, which
were previously handled by entries in the `_routing` collection are now handled
from the responsible REST action handler directly

Existing `_routing` collections will not be touched as they may contain other
entries as well, and will continue to work.