Revert to reading Plan for Index creation
When creating an index we read the entry for its collection from the plan
at the beginning again to make sure that the collection was not changed once
we want to commit changes back to plan.

We also take the opportunitiy to revert the code that checked existence of
an index with the same name, or the same index, to using the entry in plan.