downgrade WARN messages to INFO level
because the failed operations will be retried and likely succeed later on.
Additionally, fix a compile warning for a potentially uninitialized
value in FunctionUtils.h. Finally, don't abort the server process if
the batch ping timer cannot be requeued in initial replication, as this
is not a critical error. Replication can still be retried, and even has
a chance to succeed without the batch ping timer being requeued.