fix startup when using multiple coordinators.
in case multiple coordinators were used, the one that won the cluster
bootstrap race created the "root" user and was fine afterwards. other
coordinators however were not notified of the new user, and were denying
all incoming requests with an HTTP 401 "not allowed to execute this
request" response.
the change in this PR triggers an increase of the UserVersion counter
in the agency, so that other coordinators eventually get aware of the
root user being added.