when starting the RocksDB engine, first create small-sized WAL files
this is a security precaution so that when repeatedly trying to start
an instance, an instant instance failure will not lead to the disk
filling up so quickly with WAL file data.

the WAL file size is increased later on in the startup sequence, so
everything should be fine if the startup works.

this will fix a problem with the disk filling up quickly when the
arangodb starter tries to start an instance 100 times in a row but
instantaneously gets a failure back from it.