disable warning about `--server.jwt-secret`
this change hides the warning message about `--server.jwt-secret` being
insecure in case the end user has not set the option at all.
previously the warning was displayed also when not specifying the
option.
the problem was caused by not checking whether the end user had set the
option, but by checking an internal string that was filled also in other
cases then when using the `--server.jwt-secret` option.