Single validator only.
I would prefer to keep more flexibility even if that means slightly
more complexity for the customers. Multiple validators would allow
us to do better comparisons, write tests in an easier way and
enable customers to choose between fast and feature rich validation
options. I think it is very unlikely that we will be able to provide
a single solution that fits everyone.