Fix traversal executor when no register is mapped
A recent optimisation leads to cases in which the TraversalExecutor does
not get any registers mapped.

The first effect of this was an assertion being triggered that ensured
that the register mapping is not empty.

Removing this assertion does not fix the problem by itself though: we still
have to make sure we produce output even if we are not asked for any in
registers.

We achieve that by copying the inputRow.