Changed behavior for creating new collections in OneShard databases
(i.e. databases with "sharding" attribute set to "single"):

Previously it was allowed to override "distributeShardsLike" and "numberOfShards"
for each new collection. The default values were "_graphs" and 1 and were not
modified if the user did not alter them, but it was still possible to alter them.
This is now (silenly) ignored. Any attempt to set any value for
"distributeShardsLike" or "numberOfShards" for new collections in a OneShard
database will silently be ignored. The collection will automatically be sharded
like the sharding prototype and will have a single shard.

The behavior has been adjusted to match the behavior when creating collections
in a cluster started with `--cluster.force-one-shard true` option. Here any user-
supplied values for "distributeShardsLike" or "numberOfShards" were always ignored.

Now the behavior is identical for OneShard databases and databases in cluster
started with `--cluster.force-one-shard true`.

The web interface now also hides the "Distribute shards like" settings in this
case, and makes the "Number of shards" input box read-only.